refac: UPDATE MainContents.tsx, Tabs.tsx
MainContents에 있는 eventHandler, states, products_home.json을 Tabs로 분리 후 Tabs에서 코드 분리 작업 중
