+++ src/components/MainContents/MainContents.tsx
@@ -7,25 +7,26 @@ import promotionReview from '/assets/promotion/promotion-review.png';
 import productsOrigin from '../../data/json/products_home.json';
 // components
 // ! Tabs로 이동
-// import ProductCard from '../ProductCard/ProductCard';
-import Tabs from '../Tabs/Tabs';
+import ProductCard from '../ProductCard/ProductCard';
+// import Tabs from '../Tabs/Tabs';
+import Tabs, { TabPane } from '../Tabs/Tabs';
 // styles
 import './MainContents.scss';
 
 const MainContents = () => {
-    // ? 추후 products에 랭킹을 매기고 상위 랭킹 데이터만 뽑을 수 있게 해보자
+    // // ? 추후 products에 랭킹을 매기고 상위 랭킹 데이터만 뽑을 수 있게 해보자
     // limited data
     const products = productsOrigin.slice(0, 8);
 
-    const [isTabBoxOn, setIsTabBoxOn] = useState(1);
-    const [isTabTitleOn, setIsTabTitleOn] = useState(1);
+    // const [isTabBoxOn, setIsTabBoxOn] = useState(1);
+    // const [isTabTitleOn, setIsTabTitleOn] = useState(1);
 
-    const handleClick = (e, index) => {
-        // a 태그에 #으로 페이지 첫화면 상단으로 가는 것을 막자
-        e.preventDefault();
-        setIsTabTitleOn(index);
-        setIsTabBoxOn(index);
-    };
+    // const handleClick = (e, index) => {
+    //     // a 태그에 #으로 페이지 첫화면 상단으로 가는 것을 막자
+    //     e.preventDefault();
+    //     setIsTabTitleOn(index);
+    //     setIsTabBoxOn(index);
+    // };
 
     return (
         <div className="main-contents">
@@ -35,63 +36,21 @@ const MainContents = () => {
                         <h3>SEASON TREND</h3>
                     </div>
                     <div className="goods-list-content">
-                        {/* <div className="goods-tab">
-                            <div className="goods-tab-title">
-                                <ul>
-                                    <li>
-                                        <a
-                                            href="#"
-                                            className={
-                                                1 === isTabTitleOn ? 'on' : ''
-                                            }
-                                            onClick={e => handleClick(e, 1)}
-                                        >
-                                            <span>ALL</span>
-                                        </a>
-                                    </li>
-                                    <li>
-                                        <a
-                                            href="#"
-                                            className={
-                                                2 === isTabTitleOn ? 'on' : ''
-                                            }
-                                            onClick={e => handleClick(e, 2)}
-                                        >
-                                            <span>MEN</span>
-                                        </a>
-                                    </li>
-                                    <li>
-                                        <a
-                                            href="#"
-                                            className={
-                                                3 === isTabTitleOn ? 'on' : ''
-                                            }
-                                            onClick={e => handleClick(e, 3)}
-                                        >
-                                            <span>WOMEN</span>
-                                        </a>
-                                    </li>
-                                    <li>
-                                        <a
-                                            href="#"
-                                            className={
-                                                4 === isTabTitleOn ? 'on' : ''
-                                            }
-                                            onClick={e => handleClick(e, 4)}
-                                        >
-                                            <span>ACC</span>
-                                        </a>
-                                    </li>
-                                </ul>
-                            </div>
-                            <div className="goods-tab-content">
+                        {/* <Tabs
+                            products={products}
+                            isTabBoxOn={isTabBoxOn}
+                            isTabTitleOn={isTabTitleOn}
+                            onTabClick={handleClick}
+                        /> */}
+                        <Tabs>
+                            <TabPane title="ALL">
                                 <div
                                     className={
                                         1 === isTabBoxOn
                                             ? 'goods-tab-box on'
                                             : 'goods-tab-box'
                                     }
-                                    onClick={e => handleClick(e, 1)}
+                                    onClick={e => onTabClick(e, 1)}
                                 >
                                     <ul>
                                         {products.map((product, index) => (
@@ -102,13 +61,15 @@ const MainContents = () => {
                                         ))}
                                     </ul>
                                 </div>
+                            </TabPane>
+                            <TabPane title="MEN">
                                 <div
                                     className={
                                         2 === isTabBoxOn
                                             ? 'goods-tab-box on'
                                             : 'goods-tab-box'
                                     }
-                                    onClick={e => handleClick(e, 2)}
+                                    onClick={e => onTabClick(e, 2)}
                                 >
                                     <ul>
                                         {products.map((product, index) => (
@@ -119,13 +80,15 @@ const MainContents = () => {
                                         ))}
                                     </ul>
                                 </div>
+                            </TabPane>
+                            <TabPane title="WOMEN">
                                 <div
                                     className={
                                         3 === isTabBoxOn
                                             ? 'goods-tab-box on'
                                             : 'goods-tab-box'
                                     }
-                                    onClick={e => handleClick(e, 3)}
+                                    onClick={e => onTabClick(e, 3)}
                                 >
                                     <ul>
                                         {products.map((product, index) => (
@@ -136,13 +99,15 @@ const MainContents = () => {
                                         ))}
                                     </ul>
                                 </div>
+                            </TabPane>
+                            <TabPane title="ACC">
                                 <div
                                     className={
                                         4 === isTabBoxOn
                                             ? 'goods-tab-box on'
                                             : 'goods-tab-box'
                                     }
-                                    onClick={e => handleClick(e, 4)}
+                                    onClick={e => onTabClick(e, 4)}
                                 >
                                     <ul>
                                         {products.map((product, index) => (
@@ -153,14 +118,8 @@ const MainContents = () => {
                                         ))}
                                     </ul>
                                 </div>
-                            </div>
-                        </div> */}
-                        <Tabs
-                            products={products}
-                            isTabBoxOn={isTabBoxOn}
-                            isTabTitleOn={isTabTitleOn}
-                            onTabClick={handleClick}
-                        />
+                            </TabPane>
+                        </Tabs>
                     </div>
                 </div>
                 <div className="btn-goods-down-more"></div>
